Fix #486 mudando botôes
