+++ __tests__/Components/SignUpButton.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import SignUpButton from '../../src/Components/SignUpButton';
+import IconButton from '../../src/Components/IconButton';
 import Adapter from 'enzyme-adapter-react-16';
 import {shallow} from 'enzyme';
 import Enzyme from 'enzyme';
@@ -10,13 +10,13 @@ Enzyme.configure({adapter: new Adapter()});
 
 it('renders correctly', () => {
   const tree = renderer
-  .create(<SignUpButton />)
+  .create(<IconButton />)
   .toJSON();
   expect(tree).toMatchSnapshot();
 });
 it('Testing onPress', () => {
   const onPressMock = jest.fn();
-  const wrapper = shallow(<SignUpButton onPress = {onPressMock}/>);
+  const wrapper = shallow(<IconButton onPress = {onPressMock}/>);
   const button = wrapper.find('Styled(Button)').at(0);
   button.simulate('press');
   expect(onPressMock).toBeCalled();
