+++ src/Screens/EditScreen.js
@@ -7,7 +7,7 @@ import ValidationComponent from 'react-native-form-validator';
 //import axios from 'axios';
 import {Container} from 'native-base';
 import t from '../Components/Form';
-import SignUpButton from '../Components/SignUpButton';
+import IconButton from '../Components/IconButton';
 import SideBar from '../Components/SideBar';
 
 
@@ -50,10 +50,9 @@ export default class EditScreen extends ValidationComponent {
               onChange={(v) => this.onChange(v)}
             />
           </View>
-          <SignUpButton
+          <IconButton
             text = 'Salvar'
             onPress = {() => {}}
-            style = {styles.button}
             icon= 'save'
           />
         </Container>
